Sistemato il tipo di PrelQty da int a double

+++ apiPB/Dto/Models/PrelMatDto.cs
@@ -55,7 +55,7 @@ namespace apiPB.Dto.Models
 
         public string? Wc { get; set; }
 
-        public int? PrelQty { get; set; }
+        public double? PrelQty { get; set; }
 
         public bool Imported { get; set; }
